Samler all database-impl ett sted
Går bort fra tanken om å splitte opp database-impl pr subdomene, og samler det i en pakke; db
Det gjør spesielt testing av db enklere, ved at vi kan teste èn pakke for å teste all DB-integrasjon.
Beholder samtidig interfacene som har blitt etablert sammen med klientene av DB-koden, slik at avhengighetene går "riktig" vei. DB-pakken har en avhengighet til domenet - ikke motsatt.
Det gjør samtidig at core domain ikke trenger å ha et forhold til databasen i samme grad som tidligere.

Det bør videre være muligheter for å forenkle test-config enda mer, da vi kan kjøre opp hele appen med fake/mock av de ulike endepunktene og in-memory database. Vi trenger ikke like mye custom Spring Beans som det er i dag.
