+++ src/main/java/no/nav/fo/veilarbregistrering/db/ManuellRegistreringMapper.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.registrering.manuell.db;
+package no.nav.fo.veilarbregistrering.db;
 
 import lombok.SneakyThrows;
 import no.nav.fo.veilarbregistrering.registrering.BrukerRegistreringType;
