+++ src/main/java/no/nav/fo/veilarbregistrering/db/ManuellRegistreringRepositoryImpl.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.registrering.manuell.db;
+package no.nav.fo.veilarbregistrering.db;
 
 import no.nav.fo.veilarbregistrering.registrering.BrukerRegistreringType;
 import no.nav.fo.veilarbregistrering.registrering.manuell.ManuellRegistrering;
