+++ src/main/java/no/nav/fo/veilarbregistrering/db/ProfileringMapper.java
@@ -1,4 +1,4 @@
-package no.nav.fo.veilarbregistrering.profilering.db;
+package no.nav.fo.veilarbregistrering.db;
 
 import lombok.SneakyThrows;
 import no.nav.fo.veilarbregistrering.profilering.Innsatsgruppe;
