+++ src/main/java/no/nav/fo/veilarbregistrering/db/RepositoryConfig.java
@@ -1,0 +1,35 @@
+package no.nav.fo.veilarbregistrering.db;
+
+import no.nav.fo.veilarbregistrering.oppgave.OppgaveRepository;
+import no.nav.fo.veilarbregistrering.oppgave.db.OppgaveRepositoryImpl;
+import no.nav.fo.veilarbregistrering.profilering.ProfileringRepository;
+import no.nav.fo.veilarbregistrering.registrering.bruker.BrukerRegistreringRepository;
+import no.nav.fo.veilarbregistrering.registrering.manuell.ManuellRegistreringRepository;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.core.JdbcTemplate;
+
+@Configuration
+public class RepositoryConfig {
+
+    @Bean
+    OppgaveRepository oppgaveRepository(JdbcTemplate jdbcTemplate) {
+        return new OppgaveRepositoryImpl(jdbcTemplate);
+    }
+
+    @Bean
+    BrukerRegistreringRepository brukerRegistreringRepository(JdbcTemplate db) {
+        return new BrukerRegistreringRepositoryImpl(db);
+    }
+
+    @Bean
+    ManuellRegistreringRepository manuellRegistreringRepository(JdbcTemplate db) {
+        return new ManuellRegistreringRepositoryImpl(db);
+    }
+
+    @Bean
+    ProfileringRepository profileringRepository(JdbcTemplate db) {
+        return new ProfileringRepositoryImpl(db);
+    }
+
+}
