+++ src/test/java/no/nav/fo/veilarbregistrering/db/ProfileringRepositoryTest.java
@@ -1,8 +1,8 @@
-package no.nav.fo.veilarbregistrering.profilering.db;
+package no.nav.fo.veilarbregistrering.db;
 
 import no.nav.fo.veilarbregistrering.profilering.Innsatsgruppe;
 import no.nav.fo.veilarbregistrering.profilering.Profilering;
-import no.nav.fo.veilarbregistrering.profilering.db.ProfileringRepositoryImpl;
+import no.nav.fo.veilarbregistrering.db.ProfileringRepositoryImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.jdbc.core.JdbcTemplate;
